+++ cmd/contains.go
@@ -5,7 +5,7 @@ import (
 	"net"
 	"os"
 
-	"github.com/bschaatsbergen/cidr/pkg/core"
+	"github.com/bschaatsbergen/cidr/internal/core"
 	"github.com/spf13/cobra"
 )
