+++ cmd/divide.go
@@ -5,8 +5,8 @@ import (
 	"net"
 	"strconv"
 
-	"github.com/bschaatsbergen/cidr/pkg/core"
-	"github.com/bschaatsbergen/cidr/pkg/helper"
+	"github.com/bschaatsbergen/cidr/internal/core"
+	"github.com/bschaatsbergen/cidr/internal/helper"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 )
