+++ cmd/explain.go
@@ -5,8 +5,8 @@ import (
 	"net"
 	"os"
 
-	"github.com/bschaatsbergen/cidr/pkg/core"
-	"github.com/bschaatsbergen/cidr/pkg/helper"
+	"github.com/bschaatsbergen/cidr/internal/core"
+	"github.com/bschaatsbergen/cidr/internal/helper"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 )
