+++ internal/core/const.go
