+++ internal/helper/network.go
